pyproject.toml
[tool.poetry]
name = "goit-web-hw-14"
version = "0.1.14"
description = ""
authors = ["InvisUA <xndell001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
uvicorn = "^0.32.0"
alembic = "^1.14.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.9.2"}
libgravatar = "^1.0.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.17"
fastapi-mail = "^1.4.1"
bcrypt = "^4.2.1"
python-dotenv = "^1.0.1"
redis = "^5.2.0"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"